fix: Prepare for iOS with cloud builds fails on non macOS
In cloud builds, when we try to prepare the project for iOS, sometimes we fail as the `cleanDestinationApp` does not pass correct arguments to `ensurePlatformInstalled` method.
When we have cloud builds, the last argument should declare that the native prepartion should be skipped. When we do not pass it, the logic determines that the platform is not fully added and tries to add it again.
The whole `platforms/ios` directory is deleted and then we try preparing the project again. At some point the code fails as there are missing files in the `platforms/ios` (we have deleted them).

Pass the correct arguments, so the prepare will skip the native part in the mentioned case.